Lab_Project/requirements.txt
numpy==1.24.0
Pillow==9.3.0
pygame==2.1.2
PyOpenGL==3.1.6
